Add wrapper types
