+++ azure-kinect/src/lib.rs
@@ -1,15 +1,15 @@
 pub use calibration::Calibration;
 pub use camera::Camera;
 pub use capture::Capture;
-pub use enums::{ColorResolution, DepthMode};
+pub use enums::*;
 pub use device::Device;
 pub use error::Error;
 pub use factory::{DebugMessageHandler, Factory, FactoryRecord};
 pub use image::Image;
 pub use imu::Imu;
-pub use structs::{Dimension, Range};
+pub use structs::*;
 pub use transformation::Transformation;
-pub use vectors::{Float2, Float3};
+pub use vectors::*;
 
 pub mod calibration;
 pub mod camera;
@@ -18,7 +18,6 @@ pub mod device;
 pub mod enums;
 pub mod error;
 pub mod factory;
-pub mod fps;
 pub mod image;
 pub mod imu;
 pub mod playback;
